+++ src/app/modules/dataset-upload-module/upload-steps/index.tsx
@@ -3,13 +3,13 @@ import axios from "axios";
 import Container from "@material-ui/core/Container";
 import { useStoreActions, useStoreState } from "app/state/store/hooks";
 import { useChartsRawData } from "app/hooks/useChartsRawData";
-import { stepcss } from "app/fragments/datasets-fragment/style";
+import { stepcss } from "app/modules/dataset-upload-module/style";
 import { PageTopSpacer } from "app/modules/common/page-top-spacer";
 import { useAuth0, withAuthenticationRequired } from "@auth0/auth0-react";
-import MetaData from "app/fragments/datasets-fragment/upload-steps/metaData";
-import Processing from "app/fragments/datasets-fragment/upload-steps/processing";
-import FinishedFragment from "app/fragments/datasets-fragment/upload-steps/finishedFragment";
-import AddDatasetFragment from "app/fragments/datasets-fragment/upload-steps/addDatasetFragment";
+import MetaData from "app/modules/dataset-upload-module/upload-steps/metaData";
+import Processing from "app/modules/dataset-upload-module/upload-steps/processing";
+import FinishedFragment from "app/modules/dataset-upload-module/upload-steps/finishedFragment";
+import AddDatasetFragment from "app/modules/dataset-upload-module/upload-steps/addDatasetFragment";
 import { useRecoilState } from "recoil";
 import { loadedDatasetsAtom } from "app/state/recoil/atoms";
 import ObjectId from "app/utils/ObjectId";
