+++ src/app/modules/dataset-upload-module/upload-steps/previewFragment.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import Box from "@material-ui/core/Box";
 import { withStyles } from "@material-ui/core";
 import Snackbar from "@material-ui/core/Snackbar";
-import { DatasetDataTable } from "app/fragments/datasets-fragment/component/data-table";
-import { buttonFlexcss } from "app/fragments/datasets-fragment/component/dataParserToolBox/style";
+import { DatasetDataTable } from "app/modules/dataset-upload-module/component/data-table";
+import { buttonFlexcss } from "app/modules/dataset-upload-module/component/dataParserToolBox/style";
 
 interface Props {
   data: any[];
