+++ src/app/modules/dataset-upload-module/upload-steps/processing.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { ReactComponent as ErrorICon } from "app/fragments/datasets-fragment/assets/error-icon.svg";
+import { ReactComponent as ErrorICon } from "app/modules/dataset-upload-module/assets/error-icon.svg";
 
 export interface ProcessingMetaDataProps {
   setProcessingError: React.Dispatch<React.SetStateAction<boolean>>;
